Updated all files in src/rest/ with the new style.
