--- src/rest/Response.cpp
@@ -1,29 +1,0 @@
-/*
- * Response.cpp
- * Andrew Mikalsen
- * 6/23/20
- */
-
-#include "Response.hpp"
-
-namespace ods {
-    namespace rest {
-        Response::Response(std::unordered_multimap<std::string, std::string> headers, std::string body, int status) :
-            _headers(headers),
-            _body(body),
-            _status(status) {
-        }
-
-        const std::unordered_multimap<std::string, std::string>& Response::headers() const {
-            return _headers;
-        }
-
-        const std::string& Response::body() const {
-            return _body;
-        }
-
-        const int& Response::status() const {
-            return _status;
-        }
-    }
-}
