--- src/rest/Rest.cpp
@@ -1,14 +1,0 @@
-/*
- * Rest.cpp
- * Andrew Mikalsen
- * 6/5/20
- */
-
-#include "Rest.hpp"
-
-namespace ods {
-    namespace rest {
-        Rest::Rest() = default;
-        Rest::~Rest() = default;
-    }
-}
