+++ src/rest/rest.cpp
@@ -1,0 +1,12 @@
+/*
+ * rest.cpp
+ * Andrew Mikalsen
+ * 6/5/20
+ */
+
+#include "rest.h"
+
+namespace One_data_share {
+    Rest::Rest() = default;
+    Rest::~Rest() = default;
+}
